refactor(oid4vci): consolidate authorization modules into single module
