+++ crates/cloud-wallet-openid4vc/src/oid4vci/authorization/details.rs
