+++ crates/cloud-wallet-openid4vc/src/oid4vci/authorization/mod.rs
@@ -1,0 +1,7 @@
+mod details;
+mod request;
+mod response;
+
+pub use details::*;
+pub use request::*;
+pub use response::*;
