+++ crates/cloud-wallet-openid4vc/src/oid4vci/authorization/response.rs
