+++ crates/cloud-wallet-openid4vc/src/oid4vci/mod.rs
@@ -1,11 +1,10 @@
 //! OpenID4VCI protocol data models.
 
-pub mod authz_details;
-pub mod authz_request;
-pub mod authz_response;
+pub mod authorization;
 pub mod authz_server_metadata;
 pub mod claim_path_pointer;
 pub mod client;
+pub mod credential;
 pub mod credential_configuration;
 pub mod credential_formats;
 pub mod credential_offer;
@@ -13,7 +12,9 @@ pub mod credential_request;
 pub mod credential_response;
 pub mod error;
 pub mod issuer_metadata;
+pub mod metadata;
 pub mod notification;
 pub mod proofs;
+pub mod token;
 pub mod token_request;
 pub mod token_response;
